LC-separate-squares-i [Runtime: N/A; Memory: N/A]

+++ binary-search/separate-squares-i.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    double calculateAreaAbove(vector<vector<int>> &sq,double &y){
-        double area=0;
+    long double calculateAreaAbove(vector<vector<int>> &sq,double &y){
+        long double area=0;
         for(auto square:sq){
             int xc=square[0];
             int yc=square[1];
@@ -12,16 +12,16 @@ public:
                 area+= (len)*(yc+len-y);
             }
             else{
-                area+=(len*len);
+                area+=((long double)len*len);
             }
         }
         return area;
     }
     double separateSquares(vector<vector<int>>& squares) {
-        double total_area=0;
+        long double total_area=0;
         for(auto square:squares){
             int length=square[2];
-            total_area+=(length)*(length);
+            total_area+=((long double)length)*((long double)length);
         }
         double target=total_area/2;
         double low=0,high=2e9;
